consolidate util functions
