--- src/utils/damage/compareObjects.ts
@@ -1,11 +1,0 @@
-const compareObjects = (a: any, b: any) => {
-    const diff: any = {}
-    for (const key in a) {
-        if (a[key] !== b[key]) {
-            diff[key] = [a[key], b[key]]
-        }
-    }
-    return diff
-}
-
-export default compareObjects
