--- src/utils/damage/parseScalingValue.ts
@@ -1,7 +1,0 @@
-const parseScalingValue = (scalingValue: string) => {
-    const values = scalingValue.match(/\d+(\.\d+)?/g)
-    const parsedValues = values ? values.map((value) => parseFloat(value)) : [0]
-    return parsedValues
-}
-
-export default parseScalingValue
