+++ src/utils/index.ts
@@ -1,13 +1,13 @@
 export {
     attributeSections,
     availableIcons,
-    displayStats,
+    displayStats
 } from './attributes/attributeSections'
 export {
     getConstellationOptions,
     getLevelOptions,
     handleConstellationChange,
-    handleLevelChange,
+    handleLevelChange
 } from './character/characterOptions'
 export { default as handleBonusToggle } from './character/handleBonusToggle'
 export { default as compareElement } from './character/modal/compareElement'
@@ -15,10 +15,10 @@ export { default as elementColors } from './character/modal/elementColors'
 export { attributeKeyMap, fullBaseStats } from './damage/attributeKeyMapping'
 export { default as calculateDamage } from './damage/calculateDamage'
 export { default as convertBaseStats } from './damage/convertBaseStats'
-export { default as compareObjects } from './damage/compareObjects'
 export { default as damageFormula } from './damage/formulas/damageFormula'
 export { default as genericFormulaWithScaling } from './damage/formulas/genericFormulaWithScaling'
 export { default as genericFormulaWithoutScaling } from './damage/formulas/genericFormulaWithoutScaling'
-export { default as parseScalingValue } from './damage/parseScalingValue'
 export { default as useActiveConstellations } from './hooks/useActiveConstellations'
 export { default as useBaseStats } from './hooks/useBaseStats'
+export { compareObjects, parseScalingValue } from './utils'
+
